# Changed NN in-memory representation from fp32 to fp16