--- a/src/config.h
+++ b/src/config.h
@@ -48,6 +48,9 @@
 #define MAX_CPUS 128
 #endif
 
+// datatype : if you want to use half-precision float for OpenCL in-memory types define this.
+// #define USE_HALF
+
 /* Integer types */
 
 typedef int int32;
@@ -83,7 +86,11 @@ typedef  unsigned long long int uint64;
     #endif
 #endif
 
+#ifdef USE_HALF
 #include "half.hpp"
-using half_float::half;
+typedef half_float::half net_t;
+#else
+typedef float net_t;
+#endif
 
 #endif