This is a patch that uses half fp for the weights and the input/outputs.  They are converted immediately to fp32 once the GPU loads them, and all arithmetic is done in fp32.  This gives some (10~15%) performance improvement using a GTX1080.

Not sure if this is worth merging to master, but putting a pull request since it may worth thinking...