@Ttl 2017-11-26
I tried the changes and they seem to work. I also get 10% performance increase on GTX1050 and the strength of the program seems same as before.

What's the source for the half.hpp header? It doesn't have any license information.