@ihavnoid 2017-11-26
half.hpp is from:

http://half.sourceforge.net/

MIT License - is it okay?