@Ttl 2017-11-26
MIT license is compatible with GPL, there is no problem with it.

Can you restore the license comments to the beginning of half.hpp?