@ihavnoid 2017-11-26
Strange, I thought I didn't strip out any copyright info, but apparently it seems I did.  Will restore it and update it.