@sethtroisi 2017-11-26
On my CPU limited computer (i5-2400 + GTX 1080 Ti) I only see a 6% speedup.

Before:
44 game(s) played in 141 minutes = 192 seconds/game, 17957 moves = 471 ms/move

After:
44 game(s) played in 135 minutes = 185 seconds/game, 18433 moves = 442 ms/move