@isty2e 2017-11-26
I tried to compile this on my mac, but ended up with the following error:

> In file included from Network.cpp:33:
> ./Im2Col.h:39:11: error: cannot initialize a variable of type 'half_float::half *' with an rvalue of type 'value_type *' (aka 'float *')
>     half* data_col = output.data();

What would be the problem and the solution?