@isty2e 2017-11-26
Yes, it seems like working fine now.