@ihavnoid 2017-11-26
@sethtroisi - It is possible that it adds some overhead to CPU because this implementation does the final fp16->fp32 conversion from the CPU.  Still it is not clear if doing it from GPU will improve performance since it will increase the amount of data moving from GPU to CPU.  I am trying to see if there is any other potential optimization though.