@isty2e 2017-11-27
A 15% performance gain for me. However its Elo rating should be checked against fp32 version, I suppose?