@gcp 2017-11-27
>Not sure if this is worth merging to master, but putting a pull request since it may worth thinking...

I would consider merging it if the same source can support both float and half. The cleanest solution is likely templating a bunch of stuff on type, and then adding a USE_HALF in config.h. There's ways to use DEFINE's in OpenCL kernels that should make it workable to substitute the type in them.

But not all that easy, unfortunately.