@ihavnoid 2017-11-28
Well, I did make changes to support things with an ifdef.  Please see 74ac481.

Points to check:
- datatype naming
- coding style
- glaring bugs (hope none)

At least made sure it generates the first move properly on both modes.